Reorganize and simplify base CSS
This commit reorganizes the base `styles.scss` stylesheet which styles a lot of
the written elements of the Homebridge.io Homepage.

This CSS structure follows the [SMACSS](http://smacss.com/) architecture which
helps keep the CSS organized but also declarative and easy to edit.

This commit also simplifies a few of the class definitions to take advantage of
Sass' native indentation inheritance features which prevents repetition.